// frontend/src/ts/test/events/types.ts
export type EventLogContext = {
  targetWords: string[];
  // isTimedTest: boolean;
  mode: Config["mode"];
  mode2: ReturnType<typeof getMode2>;
  customTextLimitMode?: CustomTextLimitMode;
  customTextLimitValue?: number;
  isFunboxWithNospacePropertyActive?: boolean;
  bailedOut: boolean;
  koreanStatus: boolean;
};